major overhaul for authentication logic and UI

+++ src/components/neon.ts
@@ -5,20 +5,20 @@ import { cookies } from 'next/headers';
 
 
 
+
 const sql = neon(process.env.DATABASE_URL!);
 
 
 
 export async function verifyUser(credentials: AuthCredentials): Promise<User | null> {
   const users = await sql`
-    SELECT id, name, email FROM users
-    WHERE email = ${credentials.email} AND password = ${credentials.password}
-  `;
+    SELECT id, name, email,role FROM users
+    WHERE email = ${credentials.email} AND password = ${credentials.password}`;
 
   if (users.length === 0) return null;
 
   const user = users[0];
-  return { id: user.id as number, name: user.name as string, email: user.email as string, };
+  return { id: user.id as number, name: user.name as string, email: user.email as string, role: user.role as "admin" | "user" };
 }
 
 
@@ -26,23 +26,21 @@ export async function createUser(name: string, email: string, password: string):
   const result = await sql`
     INSERT INTO users (name, email, password)
     VALUES (${name}, ${email}, ${password})
-    RETURNING id, name, email
-  `;
+    RETURNING id, name, email, role`;
 
   const user = result[0];
-  return { id: user.id as number, name: user.name as string, email: user.email as string, };
+  return { id: user.id as number, name: user.name as string, email: user.email as string, role: "user" };
 }
 
 
 export async function findUserByEmail(email: string): Promise<User | null> {
   const users = await sql`
-    SELECT id, name, email FROM users WHERE email = ${email}
-  `;
+    SELECT id, name, email, role FROM users WHERE email = ${email}`;
 
   if (users.length === 0) return null;
 
   const user = users[0];
-  return { id: user.id as number, name: user.name as string, email: user.email as string, };
+  return { id: user.id as number, name: user.name as string, email: user.email as string, role: user.role as "admin" | "user" };
 }
 
 
@@ -52,61 +50,119 @@ export async function getSessionUser(): Promise<User | null> {
 
   if (!userId) return null;
 
-  try {
-    const users = await sql`SELECT id, name, email FROM users WHERE id = ${userId}`;
-    if (users.length === 0) return null;
-    return { id: users[0].id as number, name: users[0].name as string, email: users[0].email as string };
-  } catch {
-    return null;
-  }
+  const users = await sql`SELECT id, name, email, role FROM users WHERE id = ${userId}`;
+  if (users.length === 0) return null;
+  const user = users[0];
+  return { id: user.id as number, name: user.name as string, email: user.email as string, role: user.role as "admin" | "user" };
+
+}
+
+
+export async function getLibById(id: string | number): Promise<LibraryItem | null> {
+  const lib = await sql`SELECT * FROM libraries WHERE id = ${id}`;
+  if (lib.length === 0) return null;
+  return {
+    id: lib[0].id,
+    name: lib[0].name,
+    category: lib[0].category,
+    description: lib[0].description,
+    installCommand: lib[0].installcommand,
+    docsUrl: lib[0].docsurl,
+    isBookmarked: false,
+    personalNote: null,
+    createdBy: lib[0].created_by,
+    isProtected: lib[0].is_protected,
+    status: lib[0].status
+  };
+
 }
 
 
 
 
-
-
-
-
-
-export async function getLibs(query: string = "", sort: string = ""): Promise<LibraryItem[]> {
+export async function getLibs(query: string = "", sort: string = "", user: User | null): Promise<LibraryItem[]> {
 
   const searchPattern = `%${query}%`;
-
-  const user = await getSessionUser();
-
   let libraries;
 
-  if (user) {
-    if (sort === 'des') {
-      libraries = await sql`
+  if (user === null) {
+    switch (sort) {
+      case 'des':
+        libraries = await sql`
+        SELECT *, false AS isbookmarked, NULL AS personalnote FROM libraries
+        WHERE (name ILIKE ${searchPattern} OR description ILIKE ${searchPattern}) AND status = 'public'
+        ORDER BY id DESC`;
+        break;
+      case 'name':
+        libraries = await sql`
+        SELECT *, false AS isbookmarked, NULL AS personalnote FROM libraries
+        WHERE (name ILIKE ${searchPattern} OR description ILIKE ${searchPattern}) AND status = 'public'
+        ORDER BY LOWER(name) ASC`;
+        break;
+      default:
+        libraries = await sql`
+        SELECT *, false AS isbookmarked, NULL AS personalnote FROM libraries
+        WHERE (name ILIKE ${searchPattern} OR description ILIKE ${searchPattern}) AND status = 'public'
+        ORDER BY id ASC`;
+        break;
+    }
+  } else if (user.role === "user") {
+    switch (sort) {
+      case 'des':
+        libraries = await sql`
+        SELECT l.*, (b.id IS NOT NULL) AS isbookmarked, b.personalnote
+        FROM libraries l
+        LEFT JOIN bookmarks b ON b.library_id = l.id AND b.user_id = ${user?.id} 
+        WHERE (l.name ILIKE ${searchPattern} OR l.description ILIKE ${searchPattern}) AND (l.status = 'public' OR l.created_by = ${user.id})
+        ORDER BY l.id DESC`;
+        break;
+      case 'name':
+        libraries = await sql`
+        SELECT l.*, (b.id IS NOT NULL) AS isbookmarked, b.personalnote
+        FROM libraries l
+        LEFT JOIN bookmarks b ON b.library_id = l.id AND b.user_id = ${user?.id} 
+        WHERE (l.name ILIKE ${searchPattern} OR l.description ILIKE ${searchPattern}) AND (l.status = 'public' OR l.created_by = ${user.id})
+        ORDER BY LOWER(l.name) ASC`;
+        break;
+      default:
+        libraries = await sql`
+        SELECT l.*, (b.id IS NOT NULL) AS isbookmarked, b.personalnote
+        FROM libraries l
+        LEFT JOIN bookmarks b ON b.library_id = l.id AND b.user_id = ${user?.id} 
+        WHERE (l.name ILIKE ${searchPattern} OR l.description ILIKE ${searchPattern}) AND (l.status = 'public' OR l.created_by = ${user.id})
+        ORDER BY l.id ASC`;
+        break;
+    }
+  } else if (user.role === "admin") {
+    switch (sort) {
+      case 'des':
+        libraries = await sql`
         SELECT l.*, (b.id IS NOT NULL) AS isbookmarked, b.personalnote
         FROM libraries l
         LEFT JOIN bookmarks b ON b.library_id = l.id AND b.user_id = ${user?.id}
         WHERE l.name ILIKE ${searchPattern} OR l.description ILIKE ${searchPattern}
-        ORDER BY l.id DESC
-      `;
-    } else if (sort === 'name') {
-      libraries = await sql`
+        ORDER BY l.id DESC`;
+        break;
+      case 'name':
+        libraries = await sql`
         SELECT l.*, (b.id IS NOT NULL) AS isbookmarked, b.personalnote
         FROM libraries l
         LEFT JOIN bookmarks b ON b.library_id = l.id AND b.user_id = ${user?.id}
         WHERE l.name ILIKE ${searchPattern} OR l.description ILIKE ${searchPattern}
-        ORDER BY LOWER(l.name) ASC
-      `;
-    } else {
-      libraries = await sql`
+        ORDER BY LOWER(l.name) ASC`;
+        break;
+      default:
+        libraries = await sql`
         SELECT l.*, (b.id IS NOT NULL) AS isbookmarked, b.personalnote
         FROM libraries l
         LEFT JOIN bookmarks b ON b.library_id = l.id AND b.user_id = ${user?.id}
         WHERE l.name ILIKE ${searchPattern} OR l.description ILIKE ${searchPattern}
-        ORDER BY l.id ASC
-      `;
+        ORDER BY l.id ASC`;
+        break;
     }
   } else {
     libraries = await sql`
-      SELECT *, false AS isbookmarked, NULL AS personalnote
-      FROM libraries
+      SELECT * FROM libraries 
       WHERE name ILIKE ${searchPattern} OR description ILIKE ${searchPattern}
       ORDER BY id ASC
     `;
@@ -121,17 +177,19 @@ export async function getLibs(query: string = "", sort: string = ""): Promise<Li
     installCommand: lib.installcommand,
     docsUrl: lib.docsurl,
     isBookmarked: lib.isbookmarked,
-    personalNote: lib.personalnote
+    personalNote: lib.personalnote,
+    createdBy: lib.created_by,
+    isProtected: lib.is_protected,
+    status: lib.status
   }));
 }
 
 
 export async function createLib(data: Omit<LibraryItem, 'id'>): Promise<LibraryItem> {
   const newLib = await sql`
-    INSERT INTO libraries (name, category, description, installCommand, docsUrl)
-    VALUES (${data.name}, ${data.category}, ${data.description}, ${data.installCommand}, ${data.docsUrl})
-    RETURNING *;
-  `;
+    INSERT INTO libraries (name, category, description, installCommand, docsUrl, created_by, is_protected, status)
+    VALUES (${data.name}, ${data.category}, ${data.description}, ${data.installCommand}, ${data.docsUrl}, ${data.createdBy}, ${data.isProtected}, ${data.status})
+    RETURNING *;`;
 
   const lib = newLib[0];
   return {
@@ -142,7 +200,10 @@ export async function createLib(data: Omit<LibraryItem, 'id'>): Promise<LibraryI
     installCommand: lib.installcommand,
     docsUrl: lib.docsurl,
     isBookmarked: false,
-    personalNote: null
+    personalNote: null,
+    createdBy: lib.created_by,
+    isProtected: lib.is_protected,
+    status: lib.status
   };;
 }
 
@@ -156,10 +217,12 @@ export async function updateLib(data: LibraryItem): Promise<LibraryItem> {
       category = ${data.category},
       description = ${data.description},
       installCommand = ${data.installCommand},
-      docsUrl = ${data.docsUrl}
+      docsUrl = ${data.docsUrl},
+      created_by = ${data.createdBy},
+      is_protected = ${data.isProtected},
+      status = ${data.status}
     WHERE id = ${data.id}
-    RETURNING *;
-  `;
+    RETURNING *;`;
 
   const lib = updatedLib[0];
   return {
@@ -170,17 +233,17 @@ export async function updateLib(data: LibraryItem): Promise<LibraryItem> {
     installCommand: lib.installcommand,
     docsUrl: lib.docsurl,
     isBookmarked: false,
-    personalNote: null
+    personalNote: null,
+    createdBy: lib.created_by,
+    isProtected: lib.is_protected,
+    status: lib.status
   };
 }
 
 
 export async function removeLib(id: string | number) {
   const deletedLib = await sql`
-    DELETE FROM libraries
-    WHERE id = ${id}
-    RETURNING *;
-  `;
+    DELETE FROM libraries WHERE id = ${id} RETURNING *;`;
 
   const lib = deletedLib[0];
   return {
@@ -191,7 +254,9 @@ export async function removeLib(id: string | number) {
     installCommand: lib.installcommand,
     docsUrl: lib.docsurl,
     isBookmarked: lib.isbookmarked,
-    personalNote: lib.personalnote
+    personalNote: lib.personalnote, createdBy: lib.created_by,
+    isProtected: lib.is_protected,
+    status: lib.status
   };
 }
 
@@ -199,40 +264,29 @@ export async function removeLib(id: string | number) {
 export async function toggleBookmark(libraryId: string | number, user: User): Promise<boolean> {
 
   const existing = await sql`
-    SELECT id FROM bookmarks
-    WHERE user_id = ${user.id} AND library_id = ${libraryId}
-  `;
+    SELECT id FROM bookmarks WHERE user_id = ${user.id} AND library_id = ${libraryId}`;
 
   if (existing.length > 0) {
     await sql`
-      DELETE FROM bookmarks
-      WHERE user_id = ${user.id} AND library_id = ${libraryId}
-    `;
+      DELETE FROM bookmarks WHERE user_id = ${user.id} AND library_id = ${libraryId}`;
     return false;
   }
 
   await sql`
-    INSERT INTO bookmarks (user_id, library_id)
-    VALUES (${user.id}, ${libraryId})
-  `;
+    INSERT INTO bookmarks (user_id, library_id) VALUES (${user.id}, ${libraryId})`;
   return true;
 }
 
 
 export async function removeFromMyLib(libraryId: string | number, user: User): Promise<void> {
   await sql`
-    DELETE FROM bookmarks
-    WHERE user_id = ${user.id} AND library_id = ${libraryId}
-  `;
+    DELETE FROM bookmarks WHERE user_id = ${user.id} AND library_id = ${libraryId}`;
 }
 
 
 export async function updatePersonalNote(libraryId: string | number, personalNote: string | null, user: User): Promise<void> {
   await sql`
-    UPDATE bookmarks
-    SET personalnote = ${personalNote}
-    WHERE user_id = ${user.id} AND library_id = ${libraryId}
-  `;
+    UPDATE bookmarks SET personalnote = ${personalNote} WHERE user_id = ${user.id} AND library_id = ${libraryId}`;
 }
 
 
@@ -255,6 +309,14 @@ export async function getBookmarkedLibs(id: string | number): Promise<LibraryIte
     installCommand: lib.installcommand,
     docsUrl: lib.docsurl,
     isBookmarked: lib.isbookmarked,
-    personalNote: lib.personalnote
+    personalNote: lib.personalnote,
+    createdBy: lib.created_by,
+    isProtected: lib.is_protected,
+    status: lib.status
   }));
+}
+
+
+export async function setBadge(visibility: string, isProtected: boolean, id: string | number) {
+  await sql`UPDATE libraries SET status = ${visibility}, is_protected = ${isProtected} WHERE id = ${id}`;
 }
